Suporte a arquivo de entrada

+++ src/redes/Message.java
@@ -3,11 +3,11 @@ package redes;
 import java.io.Serializable;
 
 public class Message implements Serializable {
-    private int size;
+    private short size;
     private String msg;
     private String md5;
 
-    public Message(int size, String msg, String md5) {
+    public Message(short size, String msg, String md5) {
         this.msg = msg; this.size = size;   this.md5 = md5;
     }
